+++ DfE.FindInformationAcademiesTrusts/Pages/Index.cshtml.cs
@@ -1,11 +1,13 @@
 ﻿using DfE.FindInformationAcademiesTrusts.Pages.Shared;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace DfE.FindInformationAcademiesTrusts.Pages;
 
-public class IndexModel : PageModel, ISearchFormModel
+public class IndexModel : BasePageModel, IPageSearchFormModel
 {
-    [BindProperty(SupportsGet = true)] public string KeyWords { get; set; } = string.Empty;
-    public string InputId => "home-search";
+    public IndexModel()
+    {
+        ShowHeaderSearch = false;
+    }
+
+    public string PageSearchFormInputId => "home-search";
 }